chore: move general classes out of `src/fuzzer/analysis/typescript`
