+++ src/fuzzer/analysis/FunctionDef.test.ts
@@ -1,6 +1,6 @@
 import { FunctionRef, ArgTag } from "./Types";
 import { ArgDef } from "./ArgDef";
-import { ProgramDef } from "./ProgramDef";
+import { ProgramDef } from "./typescript/ProgramDef";
 import { makeArgDef, makeTypeRef } from "./TestUtils";
 
 const argOptions = ArgDef.getDefaultOptions();
