+++ src/fuzzer/analysis/FunctionDef.ts
