+++ src/fuzzer/compilers/CompilerWorker.ts
@@ -4,8 +4,8 @@ import {
   TypeScriptCompilerMessageToWorker,
   TypeScriptCompilerMessageFromWorker,
 } from "./Compiler";
-import { isError } from "./Util";
-import { TscCompilerError } from "./Types";
+import { isError } from "../Util";
+import { TscCompilerError } from "../Types";
 
 console.debug("CompilerWorker started");
